feat: add file descriptors
