+++ examples/2-methods.json
@@ -1,1 +1,1 @@
-{"codeAttributes":[{"clazz":"PartialEvaluatorDummy","method":"test0()I","instructions":[{"offset":0,"instruction":"iconst_0","finalVariablesBefore":["LPartialEvaluatorDummy;!"],"finalStackBefore":[]},{"offset":1,"instruction":"ireturn","finalVariablesBefore":["LPartialEvaluatorDummy;!"],"finalStackBefore":["0"],"finalTargetInstructions":[]}],"parameters":["LPartialEvaluatorDummy;!"],"blockEvaluations":[{"startOffset":0,"evaluations":[{"skipEvaluation":false,"isGeneralization":false,"evaluationCount":0,"instruction":"iconst_0","instructionOffset":0,"variablesBefore":["LPartialEvaluatorDummy;!"],"stackBefore":[]},{"skipEvaluation":false,"isGeneralization":false,"evaluationCount":0,"instruction":"ireturn","instructionOffset":1,"variablesBefore":["LPartialEvaluatorDummy;!"],"stackBefore":["0"]}],"branchEvaluationStack":[],"startVariables":["LPartialEvaluatorDummy;!"],"startStack":[]}]},{"clazz":"PartialEvaluatorDummy","method":"test1()I","instructions":[{"offset":0,"instruction":"iconst_1","finalVariablesBefore":["LPartialEvaluatorDummy;!"],"finalStackBefore":[]},{"offset":1,"instruction":"ireturn","finalVariablesBefore":["LPartialEvaluatorDummy;!"],"finalStackBefore":["1"],"finalTargetInstructions":[]}],"parameters":["LPartialEvaluatorDummy;!"],"blockEvaluations":[{"startOffset":0,"evaluations":[{"skipEvaluation":false,"isGeneralization":false,"evaluationCount":0,"instruction":"iconst_1","instructionOffset":0,"variablesBefore":["LPartialEvaluatorDummy;!"],"stackBefore":[]},{"skipEvaluation":false,"isGeneralization":false,"evaluationCount":0,"instruction":"ireturn","instructionOffset":1,"variablesBefore":["LPartialEvaluatorDummy;!"],"stackBefore":["1"]}],"branchEvaluationStack":[],"startVariables":["LPartialEvaluatorDummy;!"],"startStack":[]}]}]}
+{"codeAttributes":[{"clazz":"PartialEvaluatorDummy","method":"test0()I","instructions":[{"offset":0,"instruction":"iconst_0","finalVariablesBefore":["LPartialEvaluatorDummy;!"],"finalStackBefore":[]},{"offset":1,"instruction":"ireturn","finalVariablesBefore":["LPartialEvaluatorDummy;!"],"finalStackBefore":["0"],"finalTargetInstructions":[]}],"parameters":["LPartialEvaluatorDummy;!"],"blockEvaluations":[{"startOffset":0,"evaluations":[{"skipEvaluation":false,"isGeneralization":false,"evaluationCount":0,"instruction":"iconst_0","instructionOffset":0,"variablesBefore":["LPartialEvaluatorDummy;!"],"stackBefore":[]},{"skipEvaluation":false,"isGeneralization":false,"evaluationCount":0,"instruction":"ireturn","instructionOffset":1,"variablesBefore":["LPartialEvaluatorDummy;!"],"stackBefore":["0"]}],"branchEvaluationStack":[],"startVariables":["LPartialEvaluatorDummy;!"],"startStack":[]}]},{"clazz":"PartialEvaluatorDummy","method":"test0()V","instructions":[{"offset":0,"instruction":"iconst_1","finalVariablesBefore":["LPartialEvaluatorDummy;!"],"finalStackBefore":[]},{"offset":1,"instruction":"return","finalVariablesBefore":["LPartialEvaluatorDummy;!"],"finalStackBefore":["1"],"finalTargetInstructions":[]}],"parameters":["LPartialEvaluatorDummy;!"],"blockEvaluations":[{"startOffset":0,"evaluations":[{"skipEvaluation":false,"isGeneralization":false,"evaluationCount":0,"instruction":"iconst_1","instructionOffset":0,"variablesBefore":["LPartialEvaluatorDummy;!"],"stackBefore":[]},{"skipEvaluation":false,"isGeneralization":false,"evaluationCount":0,"instruction":"return","instructionOffset":1,"variablesBefore":["LPartialEvaluatorDummy;!"],"stackBefore":["1"]}],"branchEvaluationStack":[],"startVariables":["LPartialEvaluatorDummy;!"],"startStack":[]}]}]}
